# src/main/resources/swagger/api.yaml
openapi: 3.0.3
info:
  title: Cuisine Community API
  description: API pour une plateforme de cuisine sans pub et communautaire (MVP)
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Serveur local de développement

paths:
  /auth/register:
    post:
      summary: Inscription d'un nouvel utilisateur
      operationId: registerUser
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationDTO'
      responses:
        '201':
          description: Utilisateur créé avec succès
  /auth/login:
    post:
      summary: Connexion et récupération du jeton JWT
      operationId: loginUser
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginDTO'
      responses:
        '200':
          description: Connexion réussie, retourne le JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDTO'
  /recipes:
    get:
      summary: Lister toutes les recettes (Public)
      operationId: getAllRecipes
      tags:
        - Recipe
      responses:
        '200':
          description: Liste des recettes récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipeDTO'
    post:
      summary: Créer une nouvelle recette (Protégé)
      operationId: createRecipe
      tags:
        - Recipe
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeDTO'
      responses:
        '201':
          description: Recette créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDTO'
        '401':
          description: Non autorisé (Jeton manquant ou invalide)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RecipeDTO:
      type: object
      required:
        - title
        - ingredients
        - steps
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        title:
          type: string
          example: "Omelette aux fines herbes"
        description:
          type: string
          example: "Une recette simple et rapide pour le soir."
        cookingTimeMinutes:
          type: integer
          example: 10
        authorId:
          type: integer
          format: int64
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientDTO'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/StepDTO'

    IngredientDTO:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "Oeufs"
        quantity:
          type: number
          format: double
          example: 3
        unit:
          type: string
          example: "pièces"

    StepDTO:
      type: object
      required:
        - stepOrder
        - instruction
      properties:
        stepOrder:
          type: integer
          example: 1
        instruction:
          type: string
          example: "Battre les oeufs dans un bol."

    UserRegistrationDTO:
      type: object
      required:
        - email
        - password
        - pseudo
      properties:
        email:
          type: string
          format: email
          example: "chef@celtek.org"
        password:
          type: string
          format: password
          example: "MotDePasseSuperSecret123!"
        pseudo:
          type: string
          example: "CeltekChef"

    UserLoginDTO:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "chef@celtek.org"
        password:
          type: string
          format: password
          example: "MotDePasseSuperSecret123!"

    AuthResponseDTO:
      type: object
      properties:
        token:
          type: string
          description: "Le jeton JWT à envoyer dans les requêtes futures"